refactor(x-sells): move AddonPillow into the shared cross-sells module
The Insurances tab needs the same pillow-with-plus-badge as Home, and
feature modules cannot depend on each other. The cross-sells module
already declared both dataAddons and dataContract, so no build-file
change was needed.
